net-matrix/mautrix-facebook: add 0.4.1.20230412.
This is upstream "master" branch as of Apr 12, 2023. The "master" branch
has supported dev-python/mautrix-0.19.x for a while already, but the
latest release does not. This stops us from upgrading the python-mautrix
library for all other packages too. That's why I decided to include this
snapshot release into the repo.